+++ icedata-v2/src/main/java/biz/icecat/icedatav2/service/impl/DefaultSuppliersService.java
@@ -4,7 +4,7 @@ import biz.icecat.icedatav2.models.api.ApiSupplier;
 import biz.icecat.icedatav2.models.entity.SupplierEntity;
 import biz.icecat.icedatav2.repository.SupplierRepository;
 import biz.icecat.icedatav2.service.SuppliersService;
-import biz.icecat.icedatav2.service.converters.SupplierConverter;
+import biz.icecat.icedatav2.mapping.converters.SupplierConverter;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
